news-analyzer/requirements.txt
# Import base requirements
-r requirements-base.txt

# News analysis specific requirements
newspaper3k==0.2.8
nltk==3.8.1
transformers>=4.30.0
torch>=2.0.0
beautifulsoup4==4.12.3
aiohttp==3.9.5
lxml==4.9.3
asyncio==3.4.3
uuid==1.30
numpy>=1.24.0
pandas>=2.0.0
pika>=1.3.0
fastapi>=0.95.0
uvicorn>=0.21.0
requests>=2.28.0
python-dotenv>=1.0.0
tenacity>=8.2.0
yfinance>=0.2.54
redis>=5.0.0
cachetools>=5.3.0
ratelimit>=2.2.1
pydantic-settings>=2.0.0 